- Refactor en el repositorio de estado para solo hacer 1 petición y dejar 1 solo método útil. - Ajustes en el Detalles Solicitudes para agrupar las solicitudes por usuario. - Se agrega tasa de interés en la respuesta de las solicitudes. - Se refactoriza caso de uso de ObtenerSolicitudesPorEstado. - Se crea fabrica de estado para test del caso de uso. - Se ajusta el test del caso de uso. - Se agrega configuración para que Swagger pueda realizar peticiones con el token.

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/pragma/r2dbc/reactiverepository/solicitud/SolicitudReactiveRepository.java
@@ -11,7 +11,7 @@ import reactor.core.publisher.Mono;
 public interface SolicitudReactiveRepository extends ReactiveCrudRepository<SolicitudEntity, String>, ReactiveQueryByExampleExecutor<SolicitudEntity> {
 
     @Query("SELECT * FROM solicitud.solicitudes WHERE estado_id = :estadoId ORDER BY solicitud_id LIMIT :limit OFFSET :offset")
-    Flux<Solicitud> findByEstadoIdWithPagination(Integer estadoId, Integer limit, Integer offset);
+    Flux<Solicitud> findByEstadoIdWithPagination(Long estadoId, Integer limit, Integer offset);
 
     @Query("SELECT COUNT(*) FROM solicitud.solicitudes WHERE estado_id = :estadoId")
     Mono<Long> contarSolicitudesPorEstado(Integer estadoId);
